fix: reorganizing the component folders
