+++ components/organisms/company/Company.jsx
