+++ components/organisms/company/company.style.js
