+++ components/organisms/tabs/tabs.style.js
