+++ components/pages/nearby/nearbyjobs.jsx
@@ -4,7 +4,7 @@ import { View, Text, TouchableOpacity, ActivityIndicator } from "react-native";
 
 import styles from "./nearbyjobs.style";
 import { COLORS } from "../../../constants";
-import NearbyJobCard from "../../common/cards/nearby/nearbyjob-card";
+import NearbyJobCard from "../../molecules/cards/nearby/nearbyjob-card";
 import useFetch from "../../../hook/useFetch";
 
 const Nearbyjobs = () => {
