+++ components/pages/welcome/welcome.style.js
